Moved AbstractModel.php from App\Models to App\Contracts
