+++ src/Contracts/AbstractModel.php
@@ -2,17 +2,16 @@
 
 declare(strict_types=1);
 
-namespace App\Models;
+namespace App\Contracts;
 
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Model as EloquentModel;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 
 /**
  * Class AbstractModel
  * @package App
- * @mixin EloquentModel
+ * @mixin Model
  * @mixin EloquentBuilder
  * @mixin QueryBuilder
  *
