+++ src/Contracts/AbstractRepository.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Contracts;
 
-use App\Models\AbstractModel;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
