+++ src/Models/Car.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Models;
 
+use App\Contracts\AbstractModel;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class Car extends AbstractModel
