+++ src/Models/Team.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace App\Models;
 
+use App\Contracts\AbstractModel;
+
 class Team extends AbstractModel
 {
     // Define the table name explicitly if it's not the plural of the model name
